Consistency on type of stopcalls sequence number

+++ main/src/com/tramchester/domain/input/StopCall.java
@@ -35,7 +35,7 @@ public class StopCall {
         return callingPlatform.getId();
     }
 
-    public int getGetSequenceNumber() {
+    public byte getGetSequenceNumber() {
         return sequenceNumber;
     }
